requirements.txt
pandas>=2.0.0
pyarrow>=10.0.0
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.3.0
huggingface-hub>=0.20.0
python-dotenv>=0.19.0
